Makefile.am: Allow package to be built on debianesque systems
Using LIBS to assign actual libraries we want to link with puts
them last on the command line, which matters to the linker on
debian because it has the --as-needed flag enabled by default.

It's not a bad idea to do so, but it means symbols have to be
resolved in a particular order, which is what this patch
accomplishes (I think, because slashdot says so).

Signed-off-by: Andreas Ericsson <ae@op5.se>