Add save and load routines, as well as a b64_encode()
The b64_encode() function is there to provide a safe way for
plugins to create pathnames without having to think about the
input users might give, such as slashes or otherwise impractical
characters in identifying names for the stuff they want to monitor.

mp_load() and mp_save() are the "real" functions though, and they
always save and load binary data, which is why they have the
length parameters.

Signed-off-by: Andreas Ericsson <ae@op5.se>