utils_snmp: Make arguments match what check_snmp uses
... and also make sure help text prints what we actually handle,
so we don't croak on an argument that we claim to handle.

This involves changing '-v' (for --snmp-version) to '-P' (for
--protocol), but since these plugins aren't yet released, that's
not a great concern; Conformity and congruence is more important.

While at it, we also align the options in the helptext so that
privproto and privpasswd come after each other and make sure
the same is true for authproto and authpassword.

Change-Id: Ie2f48053374f6ab23ff94f8073ca4c957a93e5fe
Signed-off-by: Andreas Ericsson <ae@op5.se>