utils_snmp: Use our internal mp_snmp_synch_response()
It has one great benefit over the net-snmp one, in that it puts
the error message, no matter what it is, in the same place every
time, instead of forcing every caller to check both the response
variables and the return value of the call.

This makes it suitable for our internal callers to just return an
error and have them print the by-accessor-available error string
from the mp_snmp_context.

Change-Id: I77f58d8319689da22d6104a92c6cfd5663a093ba
Signed-off-by: Andreas Ericsson <ae@op5.se>