utils_snmp: Add a debugging function to dump the context to file
Where "file" is actually FILE *, so it's handy to pass in stdout or
stdin if one is so inclined.

This patch is exceptionally helpful when it comes to tracking down
some of the more pestersome transformations in a session struct that
the snmp library does on its own, or needs to have done in order to
do what a toddler could figure out on its own.

Change-Id: Iddfaa413bf67ad6840db8ab7c2a78fd7332ba46d
Signed-off-by: Andreas Ericsson <ae@op5.se>