utils_snmp: Force sane authentication options in mp_snmp_finalize_auth()
For instance, we *will* fail horribly if we've chosen to use AuthPriv
and neither Auth nor Priv passwords are specified. There are a bunch
of similar conditions that will cause us to go tits up really fast,
but we can avoid them by
a) Setting sensible defaults for the values we can guess with any
sort of reasonable accuracy.
b) Tell the user that something's missing when it is.

The error messages might need some work, but that can be handled when
we know from user feedback what would actually be helpful.

Change-Id: Ie423a1536e6490da3fe8d64621c5d8a5d4610509
Signed-off-by: Andreas Ericsson <ae@op5.se>