utils_snmp: Handle errors when opening snmp sessions
In this case, "handling" means "detect and provide an error message",
as it's really up to the plugin to determine what to do when we
fail. However, with this patch the ctx->errstr is set properly and
we never try to actually run queries with a NULL session, which we
did before and which caused crashes when we were unable to resolve
the provided hostname.

Change-Id: I4ec5211d4db55177ac6d06020a738c37cdf353ba
Signed-off-by: Andreas Ericsson <ae@op5.se>