Check return values of mp_snmp_walk()
Previously we would generate a coredump if we failed to create an snmp
session to the targeted host while doing an snmp walk (such as when
getting a bogus hostname or failing to resolve it, f.e.).

This patch adds the necessary checks, avoiding that particular
nastiness.

This fixes issue MON-8339.

Change-Id: I5e2d6dab261672f71aee5a91705bc8c326a1f6ea
Signed-off-by: Robin Hagman <robin.hagman@op5.com>