plugins: Don't stop parsing on non-option arguments
Beginning the optstring passed to getopt() and getopt_long() with
a plus-sign (+) triggers the posixly correct mode, which makes it
stop parsing arguments as soon as it finds a non-option argument.

That's not what we want, since it makes error handling a lot more
difficult, so get rid of it. This makes it possible to detect
out-of-place arguments and break out early, but we don't do that
in this particular patch even though it should be done.

Change-Id: Iaab9743e10024ae15f48e1b6f84d19a16a9634bc
Signed-off-by: Andreas Ericsson <ae@op5.se>