Enable a bunch of compiler warnings, along with -Werror
gcc is very good at coding C. So good that it sometimes spots
bugs before they ever go to a customer. Everything it can warn
about is worth investigating and making a conscious choice as
to whether the warning should be ignored (with a -Wno-error=foo
flag while building, which can be checked for by configure) or
fixed (which should be the most common solution).

This enables most of the warnings we have in place for other
C-based projects (ie, most of them). Let's hope we never see
any.

Change-Id: I7f2b02fce12afe1b509cee339aa19c7005875762
Signed-off-by: Andreas Ericsson <ae@op5.se>