Disallow SNMP v3 arguments when v2c or v1 is specified
Providing v3 arguments to a v1 or v2c run is certainly a user
error, so let's catch it. Previously, we would silently swap
to using SNMP v3 if any v3 arguments were present, even if
the user specifically ordered us to use a different protocol
version. Now we instead bail out when that happens.

Part of resolving MON-8356.

Change-Id: I75142e2346941bd2d6900baf4184aa61f6260a6c
Signed-off-by: Andreas Ericsson <ae@op5.se>