utils_snmp: Provide and use a function to print snmp version name
The macros for snmp versions are confusing when displayed in
debug output (v1 is 0, v2c is 1, v2u (unimplemented) is 2 and
v3 is 3), so let's provide a function printing the full name.
We have to return a const string because "2c" is not the same
as "2" for anyone anal enough to read discarded draft proposals,
and it costs us very little to be helpful with 100% accuracy.

Change-Id: I60c9d709f5650731a27a6e5d06fa4a242ea2adca
Signed-off-by: Andreas Ericsson <ae@op5.se>