Merge types, fix output and perfdata
Only one check-type "storage_used" is needed if the output is showed in
human readable bytes as well as the percentage used. Warning/critical
values can be specified with percent or bytes with prefixes [bkmgtpezy].

The perfdata always use bytes as the uom. This fixes issue MON-8391 and
issue MON-8393.

Change-Id: I095891ab59930978172e61095622058393b35cc6
Signed-off-by: Robin Hagman <robin.hagman@op5.com>