check_snmp_memory: Merge types, fix output/perf
The used memory gets calculated to not include the cached or
buffered RAM. Remove all check-types except for ram_used and
swap_used since the new output shows all of these values in the
perfdata.

Add the new option -m which can be used to change the uom for
warning/critical thresholds.

While at it, re-factor the code to make it easier to implement
the possibility to check threshold values against some
or all the performance data values in the future.

This fixes issue MON-8392 and the memory part of MON-8393.

Change-Id: Ie79ec080029fb492755051386a24d9f87e4a9788
Signed-off-by: Robin Hagman <robin.hagman@op5.com>