check_snmp_load_avg: Default option and thresholds
When specifying multiple thresholds it should be possible to set all,
parts or none of them (for example -w1,2,3 or -w,,3 or no -w at all).
The default behavior for the load check should be to return OK if no
thresholds are given.

Add tests and update old ones for this behavior, this fixes parts of
issue MON-8356.

Change-Id: I4539fe45f14ad13b5178b2fbf4ac2dcac6e998c4
Signed-off-by: Robin Hagman <robin.hagman@op5.com>