Check_snmp_load_avg: Remove 3 load options
Since the option load checks everything load1, 5 and 15 does, we remove
the options and the code associated with those checks. Remove tests
associated with those options and add a new default check which is load
instead of load1.

While at it, re-factor the code.

This is part of fixing issue MON-8356.

Change-Id: I1eb74693c5383ee7a4656c8b95ffbc4fca6d278c
Signed-off-by: Robin Hagman <robin.hagman@op5.com>