check_snmp_cpu: Load avg checks to a new plugin
check_snmp_load_avg checks the system load average and the plugin
check_snmp_cpu should be modified as described in MON-8394 to only check
the actual cpu usage. This commit adds the new plugin, modifies the name
of the types, changes the output to a better human readable text and
adds tests for this new plugin.

This fixes issue MON-8395.

Change-Id: I85df75352e595cb1bc84f33452fe8952c9478c1f
Signed-off-by: Robin Hagman <robin.hagman@op5.com>