Check_snmp_disk: Exit if we don't get the data
Check that we really get the data we asked for and exit if we don't.
Change the output of listed disks to check to not include the index
since it is just confusing since the user should not have to care about
it.

While at it do a minor clean up.

This fixes parts of issue MON-8356.

Change-Id: If4395336da9c3e0fe015e29ae3f2bd58fcab8bb8
Signed-off-by: Robin Hagman <robin.hagman@op5.com>