Check_snmp_procs: Exit if we don't get the data
We cant always trust the return value of mp_snmp_walk() since it will return 0
when it reaches the end of the oid walk. When we're not authorized to walk a
single iteration, we've gone through all iterations and get 0 back from
mp_snmp_walk(), although we were never able to do anything. Hence the global
variable.

This fixes parts of issue MON-8356.

Change-Id: I1ac94abcb458692f1777aad819abcd0f5511f929
Signed-off-by: Robin Hagman <robin.hagman@op5.com>