Check commands for the new SNMP plugins
Modify the meta-data-files to use the new check_snmp_plugins check
commands. Remove check_snmp_cpu from the makefile and spec-file since we
split it into check_snmp_load_avg and the parts left in check_snmp_cpu
is not ready yet.

This fixes the meta-data part of issue MON-8494.

Change-Id: I5fcd9de3924e39998f5df974311b77f989d7def8
Signed-off-by: Robin Hagman <robin.hagman@op5.com>