Add the tap library and a small example for check_snmp_cpu
While we primarily use libcheck to run unit tests on C code,
the upstream project where we'd like to put these plugins
use tap exclusively, so that's what we're sticking with.

While at it, we remove the phpunit tests from running during
'make check', since they don't work anywhere but on the
CI servers anyway, and also add 'make check' build step in
the spec file.

We prudently maintain .gitignore in good working order
as well, and also add a few more automake products to it.

This resolves MON-8604.

Change-Id: I40beb3e9098354db4133401b8fd2c55e506af2ac
Signed-off-by: Andreas Ericsson <ae@op5.se>