Add filtering capabilities to check_snmp_disk
I may have gone a little bit overboard on this one, as the
original issue only mentioned regular expressions. However,
the storage units we need to filter out to be able to run a
somewhat sensible "check all disks" thing are of the sort
that regular expressions don't really cut it.

Now it's perfectly possible (and even recommended, as we
do it internally) to add multiple filters and with great
precision select which storage units one wants to monitor,
like so:
check_snmp_disk --exclude-size -100M --exclude-size +200G \
    --exclude-regex ^/run --exclude-regex ^/sys/fs \
	--exclude-blocksize =1k --types fixeddisk,networkdisk

which still doesn't include a threshold anywhere in there.
I've sort of forgotten to add that functionality back in,
so this plugin is currently a really elaborate filtering
mechanism that wins out over grep only because it knows
how to compare numbers.

This sort of fixes MON-8396, though I'm sure I'll get a
slew of feedback on how this is implemented. I'm particularly
ashamed of some of the function names, but I couldn't for
the life of me think of any better.

Change-Id: Ief0a183b560efd3f5249879649351fcfea593572
Signed-off-by: Andreas Ericsson <ae@op5.se>