check_snmp_disk: Re-add threshold matching
It's sort of necessary for a plugin to work properly, so let's
make sure we do it.

Barring tests, which I won't start adding this late on a friday,
everything is now complete for issue MON-8396 and MON-8610,
although I must say the output text could do with some additional
work. I'm really bad at that sort of thing though, so I'll wait
until monday to get some feedback on that.

While at it, we auto-ignore disks without description. They can't
sensibly be selected and we can't sensibly tell the user which
disk it is. This prevents a coredump in the unlikely case we
get storage unit info but are somehow prevented from reading the
storage unit description. Very, very weird snmp settings or faulty
testcases can cause this to happen, so let's see to it that we
avoid a coredump at least.

We also update the tests to match the new behaviour.

Change-Id: Id388ce222c286532be18595343a45f11d8f5cf69
Signed-off-by: Andreas Ericsson <ae@op5.se>