check_snmp_memory: handle overflow values
If the checked host has a bit more than 4GiB of RAM the int overflows.
Change the variables handling the response from SNMP to unsigned int
since that is what SNMP gives us. Also typecast the calculations to
double to avoid future overflows.

Since we change from int to unsigned int, we can no longer set the
values to -1 and check that it changes to verify that we got data with
out SNMP response. Therefore add variable that becomes set if we get
some data from the request.

This fixes issue MON-8645.

Change-Id: I27dc21c6a9f802c21d5a58618aa148489e03e623
Signed-off-by: Robin Hagman <robin.hagman@op5.com>