Remove dependency on naemon
Only rbtree from naemon was really used, which is just to copy to this
project (it has a really open license), and thus, we can remove the
dependency to naemon.

Also bitmap_set, bitmap_isset and bitmap_destroy was called from a
plugin, but since the bitmask variable only was set to NULL, and it
wasn't really clear what it should do, it only reetruned false, and
didn't do anything. So that part is just removed.

Thus, make those plugins not dependent on naemon, so they can be used
for anything else, like sensu, icinga, schinken och nagios too

Resolves: MON-8676

Change-Id: I3955be11e3aef03ea9ed3cf3754dbf5e6c6da2c4
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>