check_snmp_disk: fix failed tests and warnings
A few tests failed before but did not trigger a failed build but now
they do. Fix the bug that the test found where strncasecmp() compared
the first n characters of the string "ram" which matches "ramdisk", we
now compare the first n characters of "ramdisk" instead which does not
match "ram". While at it solve a few warnings.

This commit is not part of an issue but relates to fixes for MON-8676.

Change-Id: If1cc4d63067781206969a8ddf6643f279e224a17
Signed-off-by: Robin Hagman <robin.hagman@op5.com>