Rename check_by_snmp_load_custom to include avg
The old command for the custom check did not include the _avg for some
reason. This is confusing when we have another plugins called
check_snmp_load. Rename the command while we are in the process of
renaming the plugins. I have already renamed the check in the default
configuration.

This is part of fixing issue MON-8758.

Change-Id: Ica632634bd87c357740c0834f1a88a6182ce5d8f
Signed-off-by: Robin Hagman <robin.hagman@op5.com>