check_by_snmp_cpu: Add new dir and graph template
We would like to have nice graphs with the plugins so add a new
directory in our repo called pnp and store the graph templates there.

Add the template check_by_snmp_cpu.php to the dir and modify the
spec-file to place the graph template in the right place with correct
permissions during installation.

The colors used in the graph is temporary and will probably change after
a UX designer takes a look at it but this is an OK first approximation
of the new graph template.

This fixes issue MON-8760.

Change-Id: Ib612f4baf3209f1e11a13f835b008b5122e236ef
Signed-off-by: Robin Hagman <robin.hagman@op5.com>