check_by_snmp_memory: Add graph template
Add the graph template with the same colors that check_linux_stat uses.

The issue MON-8572 states that the graph should show "A stacked graph
showing memused, memfree, memcached", since the plugin reports buffered
as well we include it. Memfree includes chached, buffered and the rest
(could be called the free memory) so the
graph looks weird when included. Therefore show memused, membuffered and
memchached in the second graph and memused and memfree in the first.

For swap it states "A stacked graph showing swapused, swapcached,
swapfree" but our plugin only report swapused and swapfree so we only
graph that.

Since some of the colors looks incredibly bad together I hard coded
some colors to look OK.

This fixes issue MON-8761.

Change-Id: I3f62a7cc36ee4d9bf4918f690a9dbc45a11503f3
Signed-off-by: Robin Hagman <robin.hagman@op5.com>