check_by_snmp_disk: Add graph template
Should look better than the standard template. We decided not to use op5
colors since it would just add another graph style so we try to use the
same theme as check_linux_stat.

This template is a copy of the default one with small changes to color.

This fixes issue MON-8762.

Change-Id: If52b983a47e8e63d2aa2e2f5cc1bfbf57cdd49fa
Signed-off-by: Robin Hagman <robin.hagman@op5.com>