Refactor unit tests
All the unit tests had a bunch of boilerplatey code that was
nearly identical between testcases, along with the testdata
and the actual tests.

This patch consolidates the helper functions into a test_helper
class, where one can put tests such as "make sure the --version
and help arguments work as intended", because it's supposed to
run for ALL plugins, and one can do things like "make sure we
get the same values regardless of snmp version used to access
the server".

The tests for '--help' and '--version' flags are also added,
resolving the testing part of MON-8688, which was previously
manually verified.

Change-Id: I28272a7499f1f00087dafe6c83cdf2b4b600bc4b
Signed-off-by: Andreas Ericsson <ae@op5.se>