check_by_snmp_procs: Create basename from RunPath to name programs
The "Name" field is seriously unreliable as it only gives you the
first 15 bytes (or whatever it was; I forget) of the "program name",
so people will conceivably want to use the basename(first_word) of
the much larger RunPath field instead. This also solves the issue
of binaries running from a particular directory as well as some
not-so-clever child-spawning programs kicking off programs with
a kernel program name which is a concatenation of their entire
argument list. Neato.

This still doesn't make it possible to filter on other fields than
"name", but that's a different issue.

We also add tests for the first_word_basename() function, ensuring
that it works properly and is easy to write more tests for.

Change-Id: Ica6ae8ac1ef4ae35780a7acb207bca0a4b1b2600
Signed-off-by: Andreas Ericsson <ae@op5.se>