utils_snmp: Add helpers for handling ascii-based OID's
Some fields in the SNMP christmas tree are indexed by an ascii-spelled
name, which is appended to the oid in order to make it unique.

That's pretty nifty, since it means we don't have to walk() a long
list of oid's in order to get the one people know what they named,
but it does mean some minor extra complexity in terms of how such
OID's are handled. This patch adds helpers to deal with that, and
although we currently only use one of them, it's natural to add its
mirror at the same time, in order to get full test coverage of both
functions at the same time.

This is prep-work for resolving MON-8875.

Change-Id: I9a75189cfc9fe8a8f3d74a277135a2c8e61d6658
Signed-off-by: Andreas Ericsson <ae@op5.se>