Add check_by_snmp_extend
This runs plugins via SNMP by using the nsExtend tree and its
various branches. There's currently no way to list the configured
commands, but that's a separate issue (and one that can be solved
by snmpwalk'ing anyway).

This has the full functionality, although "full functionality" in this
case means querying two oid's, fetching the output, dumping it
straight out and then exiting with the result code SNMP provided us
with.

This resolves MON-8875.

Change-Id: I8fbfbebadedf604e6d76ab2985b2478b591b95d0
Signed-off-by: Andreas Ericsson <ae@op5.se>