Add check_by_snmp_disk_io
We only need to load the last state and then save the current one to
calculate the rates since it will be calculated as an averages over the
time waited. One somewhat irritating thing with this is that the time
span will be different depending on when the check was performed.

This fixes issue MON-8397.

Change-Id: Ic71b5fbad71131c6fe8c6a9e47bd545b37791ea8
Signed-off-by: Robin Hagman <robin.hagman@op5.com>