Check_by_snmp_procs: Remove -f flag
It is used if the user wants to turn of performance data. No one wanted
this feature and it should therefore be removed. This was found when we
started running the "invalid option" tests on all the check_by_snmp_*
plugins which uses the -f flag as an invalid option.

This was found when fixing MON-9042.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>