Update tests to use SNMPv3 and SNMPv2c
Test all combinations of SNMPv3 that we support, implement the use of a
dataProvider and update all test-cases that need to use it.

The check_by_snmp_* plugins doesn't support the use of context name so
do a small workaround with snmpsim. When we have the support for context
name we can easily implement the change.

The currently used localstatedir for the tested plugins are
/var/{plugin_name}, to be able to run the tests as a non-root user use
the MP_STATE_PATH to change the path to something like /tmp/. Implement
the removal of temporary files in tearDown to clean up after us.

While at it do some formating fixing with lines that are way too long.

For SNMPv3 testing: When not using a context name snmpsim takes that as
an empty string and uses a default directory to look for the data it
should respond with if the caller have the correct SNMPv3 credentials.
The default directory in our case is hard-coded when we run the SNMPv3
tests.

When we have plugins which uses context name: snmpsim uses the
md5:ed path to the .snmprec file as the name of the file. Let's say
the path is /tmp/check_by_snmp_cpu/
1e7925d84436478f547ac4684a6aa209/data/
e0b14e2b6c981ce1ac42a6b5ca466758.snmprec then, the part that should be
md5:ed is 1e7925d84436478f547ac4684a6aa209/data/ and you will get the
correct file name that the caller (plugin) should use as context name.

This fixes issue MON-8323.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>