check_by_snmp_disk_io: test missing unuseddata
Update tests to verify the issue. The output from the plugin was: "Failed to read data for storage unit 1 (ram0). Please check your SNMP configuration" which this commit verifies. Found a bug with the testing where we did not return the snmp data, fix that so that we actually checks the correct snmp OIDs.

This is part of fixing MON-9656.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>