test/test_helper.php: Replace deprecated split() with explode()
As per php.net/manual/en/function.split.php:
Warning This function was DEPRECATED in PHP 5.3.0, and REMOVED in PHP 7.0.0.

Since PHP 7 has been out for quite some time and is the default for many
potential contributors, we should support it. It also doesn't break the
tests for anyone running PHP >= 4, which should be everyone, it's a
reasonably safe change to make.

While at it, fix one block of wonky indentation.

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>