check_by_snmp_disk: Allow filters to match entries with 0
Previously we would try to filter out things with a total size of 0,
but since we did so by recycling the variable that held the actual
value we were supposed to compare to, we could never match 0,
since no such filters would ever be activated.

Solve this by adding a separate variable which we set when a filter
is supposed to be used, thus avoiding the "can't match 0" problem.

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>