check_by_snmp_disk: Fix the 'blocks' filter
This filter is kinda useless, as it's all about disk-size
divided by actual blocksize. It's simply a number people don't
normally care about. However, we pretended to support them earlier,
but didn't really, as none of the evaluation code that actually
implemented the functionality was there.

This patch adds it, thus ensuring that we support what we claim to
support.

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>