check_by_snmp_disk: Add --strip-descr-from
Some systems (notably windows machines) have exceptionally long
diskInfo.Description fields, which play havoc with PNP and also
looks very ungainly in the output.

In order to combat that problem, we add --strip-descr-from, which
can be used to specify a string (or a simple character) from where
we should strip the rest of the description string.

In order to test this with relevant data (ie, from a Windows machine)
we add a separate suite with windows-only data.

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>