check_by_snmp_disk: Add options to measure disks based on free space
The free space check is done only in actual bytes (where SSI suffixes
can be used) and is only used in case the size of the disk exceeds a
configurable threshold.

The new options introduced are --free-bytes-cutoff --free-bytes-critical
and --free-bytes-warning, each with well-defined help texts.

Since free diskspace is always a good thing, we don't bother with
ranges, but instead use the threshold values as "at least this much
should be available to not trigger an alert".

Tests for the recently added options are available in the recently
added suite "win_disk_Test.php"

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>