check_by_snmp_disk: Add option --cleanup-description
Some disklabels (notably those on Windows) sometimes contain chars
that don't sit so well with PNP. In order to provide a way to
circumvent that for those who so desire, we add the non-default option
cleanup-description, which replaces sequences of everything but
alphanumerical chars and :/_- with an underscore. If the sequence
ends the string, it is simply removed.

No behaviour is changed without applying the commandline argument,
which lacks a short option.

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>